♻️ Move CompassPeopleUtility to util package
- open to names for util package
